requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2026.1
pandas==2.2.1
openpyxl==3.1.2
